backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
yfinance
pandas
python-dotenv
